# OpenAPI.yaml
openapi: 3.0.0
info:
  title: Challange Chapter 6
  contact: {}
  version: '1.0'
servers:
- url: http://localhost:8000/v1/api
  variables: {}
paths:
...
  /registrasi-admin:
    post:
      tags:
      - Users
      summary: Register Admin
      operationId: RegisterAdmin
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/RegisterAdminRequest'
              - example:
                  name: admin
                  email: admin@gmail.com
                  password: '123456'
                  confPassword: '123456'
                  role: admin
            example:
              name: admin
              email: admin@gmail.com
              password: '123456'
              confPassword: '123456'
              role: admin
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    RegisterAdminRequest:
      title: RegisterAdminRequest
      required:
      - name
      - email
      - password
      - confPassword
      - role
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        confPassword:
          type: string
        role:
          type: string
      example:
        name: admin
        email: admin@gmail.com
        password: '123456'
        confPassword: '123456'
        role: admin
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...